refactor: reorganize component structure and update imports for clarity
