+++ crates/tss-gui/src/component/display/core_badge.rs
