+++ crates/tss-gui/src/component/display/domain_badge.rs
