+++ crates/tss-gui/src/component/display/domain_card.rs
