+++ crates/tss-gui/src/component/display/metadata_card.rs
