+++ crates/tss-gui/src/component/display/selectable_row.rs
