+++ crates/tss-gui/src/component/display/status_badge.rs
