+++ crates/tss-gui/src/component/display/status_card.rs
