+++ crates/tss-gui/src/component/feedback/mod.rs
@@ -7,12 +7,25 @@
 //! - **ProgressModal**: Progress indicators in modal
 //! - **ProgressBar**: Progress indicators
 //!
-//! # Planned Migration
+//! # Example
 //!
-//! Components to move here:
-//! - `toast.rs`
-//! - `modal.rs`
-//! - `progress_modal.rs`
-//! - `progress_bar.rs`
+//! ```ignore
+//! use tss_gui::component::feedback::{modal, ProgressBar};
+//!
+//! // Modal dialog
+//! modal("Title", content, on_close);
+//!
+//! // Progress bar
+//! ProgressBar::new(0.75)
+//!     .label("Processing...")
+//!     .view();
+//! ```
 
-// Components will be added as they are migrated
+mod modal;
+mod progress_bar;
+mod progress_modal;
+pub mod toast;
+
+pub use modal::{confirm_modal, modal};
+pub use progress_bar::ProgressBar;
+pub use progress_modal::progress_modal;
