+++ crates/tss-gui/src/component/mod.rs
@@ -1,162 +1,38 @@
 //! Reusable UI components for Trial Submission Studio.
 //!
-//! This module provides building blocks for constructing views:
-//!
-//! - **Layout**: `master_detail`, `sidebar`, `tab_bar`, `MasterPanelHeader`
-//! - **Overlays**: `modal`, `progress_modal`
-//! - **Form**: `search_box`, `form_field`, `TextField`
-//! - **Display**: `status_badge`, `data_table`, `MetadataCard`, `StatusCard`
-//! - **Feedback**: `EmptyState`, `LoadingState`, `ErrorState`
-//! - **List Items**: `VariableListItem`, `SimpleListItem`, `SelectableRow`
-//! - **Actions**: `ActionButton`, `ActionButtonList`
-//! - **Icons**: Use `iced_fonts::lucide::*` directly (see <https://lucide.dev/icons/>)
-//!
-//! # Design Philosophy
-//!
-//! Components use the builder pattern and return `Element<M>`.
-//! This provides:
-//! - Simple composition
-//! - Type-safe message passing
-//! - Easy customization per use case
+//! Components are organized into categories:
+//! - `layout`: Page structure, sidebars, split views, tab bars
+//! - `panels`: List panels, detail panels, headers, sections
+//! - `display`: Badges, cards, tables, list items
+//! - `inputs`: Search boxes, text fields, form fields
+//! - `feedback`: Toasts, modals, progress indicators
 //!
 //! # Usage
 //!
+//! Import components from their respective submodules:
+//!
 //! ```rust,ignore
-//! use tss_gui::component::{
-//!     master_detail_with_pinned_header, SearchFilterBar, VariableListItem,
-//!     DetailHeader, MetadataCard, StatusCard,
-//! };
-//!
-//! // Search + filter bar
-//! let header = SearchFilterBar::new(&search, "Search...", |s| Msg::Search(s))
-//!     .filter("Unmapped", unmapped, Msg::Toggle)
-//!     .stats("15/20 mapped")
-//!     .view();
-//!
-//! // Variable list item
-//! let item = VariableListItem::new("STUDYID", Msg::Select(0))
-//!     .label("Study Identifier")
-//!     .leading_icon(lucide::circle_check().size(12).color(SUCCESS))
-//!     .trailing_badge("Req", ERROR)
-//!     .selected(true)
-//!     .view();
+//! use tss_gui::component::layout::{SplitView, PageHeader, sidebar};
+//! use tss_gui::component::panels::{DetailPanel, SearchFilterBar};
+//! use tss_gui::component::display::{DomainCard, status_badge, EmptyState};
+//! use tss_gui::component::inputs::{TextField, search_box};
+//! use tss_gui::component::feedback::{modal, ProgressBar};
 //! ```
 
-// Allow unused exports - these are public API items that may not be used internally
-// Component library provides building blocks - not all are used yet
-#![allow(unused_imports)]
-#![allow(dead_code)]
-
-// Core modules
-mod core_badge;
-mod data_table;
-mod domain_badge;
-mod empty_state;
-mod form_field;
-mod modal;
-mod page_header;
-mod progress_modal;
-mod search_box;
-mod section_card;
-mod selectable_row;
-mod sidebar;
-mod status_badge;
-mod tab_bar;
-
-// New comprehensive layout components
-mod action_button;
-mod detail_header;
-mod domain_card;
-mod master_panel;
-mod metadata_card;
-mod progress_bar;
-mod search_filter_bar;
-mod status_card;
-mod text_field;
-pub mod toast;
-mod variable_list_item;
-
-pub use sidebar::{SidebarItem, sidebar};
-pub use tab_bar::{Tab, tab_bar};
-
-// Master panel components
-pub use master_panel::{MasterPanelHeader, MasterPanelSection, master_panel_empty};
-
-// Search and filter components
-pub use search_filter_bar::{FilterToggle, SearchFilterBar};
-
-// Overlay components
-pub use modal::{confirm_modal, modal};
-pub use progress_modal::progress_modal;
-
-// Form components
-pub use form_field::{form_field, number_field};
-pub use search_box::{search_box, search_box_compact, search_box_with_filter};
-pub use text_field::{TextAreaField, TextField};
-
-// Display components
-pub use data_table::{TableColumn, data_table};
-pub use status_badge::{
-    Status, count_badge, mapping_status_badge, status_badge, status_badge_with_icon, status_dot,
-    validation_badge,
-};
-
-// Metadata display components
-pub use metadata_card::{MetadataCard, metadata_row, metadata_row_wide};
-
-// Status card components
-pub use status_card::{
-    StatusCard, status_card_neutral, status_card_success, status_card_unmapped, status_card_warning,
-};
-
-// Feedback components (builder pattern)
-pub use empty_state::{EmptyState, ErrorState, LoadingState, NoFilteredResults};
-
-// Section components
-pub use section_card::{SectionCard, panel, status_panel};
-
-// Badge components
-pub use core_badge::{core_badge, core_badge_if_important};
-pub use domain_badge::{domain_badge, domain_badge_small};
-
-// Progress components
-pub use progress_bar::ProgressBar;
-
-// Domain card component (for home view)
-pub use domain_card::DomainCard;
-
-// List item components
-pub use selectable_row::{DomainListItem, SelectableRow};
-pub use variable_list_item::{SimpleListItem, VariableListItem};
-
-// Header components
-pub use detail_header::DetailHeader;
-pub use page_header::{PageHeader, page_header_simple};
-
-// Action button components
-pub use action_button::{ActionButton, ActionButtonList, ActionButtonStyle, edit_buttons};
-
-// Icons: Use iced_fonts::lucide directly
-// Re-export font bytes for convenience (load in main.rs)
-pub use iced_fonts::LUCIDE_FONT_BYTES;
-
-// =============================================================================
-// NEW COMPONENT ORGANIZATION (Phase 5)
-// =============================================================================
-// These subdirectories establish the new component hierarchy.
-// Components will be migrated here incrementally.
-
-/// Layout components (split views, page structure)
+/// Layout components (split views, page structure, navigation)
 pub mod layout;
 
-/// Panel components (list panels, detail panels)
+/// Panel components (list panels, detail panels, sections)
 pub mod panels;
+
+/// Display components (badges, cards, tables, list items)
+pub mod display;
 
 /// Input components (search, text fields, forms)
 pub mod inputs;
 
-/// Display components (tables, badges, empty states)
-pub mod display;
-
 /// Feedback components (toasts, modals, progress)
 pub mod feedback;
+
+// Icon font bytes for convenience
+pub use iced_fonts::LUCIDE_FONT_BYTES;
