+++ crates/tss-gui/src/component/panels/detail_header.rs
