+++ crates/tss-gui/src/component/panels/mod.rs
@@ -6,11 +6,15 @@
 //! - **ScrollableListPanel**: ListPanel wrapped in a scrollable container
 //! - **DetailPanel**: Detail view panel with header, sections, and actions
 //! - **EmptyDetailView**: Empty state for detail panels
+//! - **DetailHeader**: Header for detail views with title and actions
+//! - **MasterPanel**: Master panel components for layouts
+//! - **SectionCard**: Card sections for content grouping
+//! - **SearchFilterBar**: Search and filter controls
 //!
 //! # Example
 //!
 //! ```ignore
-//! use tss_gui::component::panels::{ListPanel, DetailPanel};
+//! use tss_gui::component::panels::{ListPanel, DetailPanel, DetailHeader};
 //!
 //! // Full-featured list panel
 //! ListPanel::new()
@@ -29,8 +33,16 @@
 //!     .view();
 //! ```
 
+mod detail_header;
 mod detail_panel;
 mod list_panel;
+mod master_panel;
+mod search_filter_bar;
+mod section_card;
 
+pub use detail_header::DetailHeader;
 pub use detail_panel::{DetailPanel, EmptyDetailView};
 pub use list_panel::{ListPanel, ScrollableListPanel};
+pub use master_panel::{MasterPanelHeader, MasterPanelSection, master_panel_empty};
+pub use search_filter_bar::{FilterToggle, SearchFilterBar};
+pub use section_card::{SectionCard, panel, status_panel};
