+++ crates/tss-gui/src/state/view_state.rs
@@ -163,7 +163,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in domain editor view.
     #[inline]
-
     pub fn current_tab(&self) -> Option<EditorTab> {
         match self {
             Self::DomainEditor { tab, .. } => Some(*tab),
@@ -173,28 +172,24 @@ impl ViewState {
 
     /// Check if currently in domain editor view.
     #[inline]
-
     pub fn is_domain_editor(&self) -> bool {
         matches!(self, Self::DomainEditor { .. })
     }
 
     /// Check if currently in home view.
     #[inline]
-
     pub fn is_home(&self) -> bool {
         matches!(self, Self::Home { .. })
     }
 
     /// Check if currently in source assignment view.
     #[inline]
-
     pub fn is_source_assignment(&self) -> bool {
         matches!(self, Self::SourceAssignment { .. })
     }
 
     /// Check if currently in export view.
     #[inline]
-
     pub fn is_export(&self) -> bool {
         matches!(self, Self::Export(_))
     }
@@ -210,7 +205,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in domain editor view.
     #[inline]
-
     pub fn mapping_ui_mut(&mut self) -> Option<&mut MappingUiState> {
         match self {
             Self::DomainEditor { mapping_ui, .. } => Some(mapping_ui),
@@ -222,7 +216,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in domain editor view.
     #[inline]
-
     pub fn mapping_ui(&self) -> Option<&MappingUiState> {
         match self {
             Self::DomainEditor { mapping_ui, .. } => Some(mapping_ui),
@@ -234,7 +227,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in domain editor view.
     #[inline]
-
     pub fn validation_ui_mut(&mut self) -> Option<&mut ValidationUiState> {
         match self {
             Self::DomainEditor { validation_ui, .. } => Some(validation_ui),
@@ -246,7 +238,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in domain editor view.
     #[inline]
-
     pub fn preview_ui_mut(&mut self) -> Option<&mut PreviewUiState> {
         match self {
             Self::DomainEditor { preview_ui, .. } => Some(preview_ui),
@@ -258,7 +249,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in domain editor view.
     #[inline]
-
     pub fn normalization_ui_mut(&mut self) -> Option<&mut NormalizationUiState> {
         match self {
             Self::DomainEditor {
@@ -272,7 +262,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in domain editor view.
     #[inline]
-
     pub fn supp_ui_mut(&mut self) -> Option<&mut SuppUiState> {
         match self {
             Self::DomainEditor { supp_ui, .. } => Some(supp_ui),
@@ -284,7 +273,6 @@ impl ViewState {
     ///
     /// Call this when data changes that would affect the preview.
     /// The preview will be rebuilt on next access.
-
     pub fn invalidate_preview(&mut self) {
         if let Self::DomainEditor { preview_cache, .. } = self {
             *preview_cache = None;
@@ -293,7 +281,6 @@ impl ViewState {
 
     /// Get the cached preview DataFrame if available.
     #[inline]
-
     pub fn preview_cache(&self) -> Option<&DataFrame> {
         match self {
             Self::DomainEditor { preview_cache, .. } => preview_cache.as_ref(),
@@ -302,7 +289,6 @@ impl ViewState {
     }
 
     /// Set the preview cache.
-
     pub fn set_preview_cache(&mut self, df: DataFrame) {
         if let Self::DomainEditor { preview_cache, .. } = self {
             *preview_cache = Some(df);
@@ -313,7 +299,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in source assignment view.
     #[inline]
-
     pub fn source_assignment_ui_mut(&mut self) -> Option<&mut SourceAssignmentUiState> {
         match self {
             Self::SourceAssignment { assignment_ui, .. } => Some(assignment_ui),
@@ -325,7 +310,6 @@ impl ViewState {
     ///
     /// Returns `None` if not in export view.
     #[inline]
-
     pub fn export_state_mut(&mut self) -> Option<&mut ExportViewState> {
         match self {
             Self::Export(state) => Some(state),
